requirements.txt
openai==0.27.0
PyPDF2==2.11.2
scikit-learn==1.2.2
python-dotenv==0.21.0
sentence-transformers==2.2.2
numpy==1.24.3